% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGene.R
\name{annotateGene}
\alias{annotateGene}
\title{annotate probes by genes}
\usage{
annotateGene(probe_GR, gene_GR, promoter_up = 2000L, promoter_down = 500L)
}
\arguments{
\item{probe_GR}{(GRanges) output of getLocation(MethylSet-Genomic)}

\item{gene_GR}{(GRanges) gene transcript region. must have name,name2}
}
\value{
(list of GRanges)
1) promoters: probe-to-promoter matches
2) genes:	probe-to-genebody matches
}
\description{
annotate probes by genes
}

