% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateSig.R
\name{annotateSig}
\alias{annotateSig}
\title{annotate foreground background probes}
\usage{
annotateSig(mset, dmpIn, annoFiles, FisherBed = NULL, qCutoff = 0.05,
  outDir, fPrefix = "", ...)
}
\arguments{
\item{mset}{(GenomicRatioSet)}

\item{dmpIn}{(DataFrame) output of minfi::dmpFinder()}

\item{annoFiles}{(data.frame) bed files for various annotations, 
with names (Type), and path to bed file (File)
Should have special entries for "Exons","Introns","Genes"}

\item{FisherBed}{(list) bed files containing genomic regions in
which overrepresentation of probes must be tested. If NULL, test is 
not performed}

\item{qCutoff}{(numeric 0-1) Q value cutoff for foreground}

\item{outDir}{(char) path to output dir}

\item{fPrefix}{(char) a name for the analysis, to add to output files
e.g. fPrefix="neuron" would create output files named 
"dmp_neuron_annotateSig_yymmdd.pdf"}

\item{...}{params for dmp_pathwayGREAT()}

\item{geneDomain}{(numeric) max distance (kb) from TSS at which a probe is considered
within the gene's domain. Used for GREAT analysis. Should be large enough to include
any probe that would be part of a gene's domain by GREAT's calculation. Should be small
enough that the resulting bed files can be uploaded to the GREAT server.
(http://bejerano.stanford.edu/great/public/html/)}
}
\description{
annotate foreground background probes
}
\details{
Runs variety of analyses to identify characteristics of 
significant probes
}

