% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgByGroup.R
\name{avgByGroup}
\alias{avgByGroup}
\title{average beta values by group}
\usage{
avgByGroup(b, g)
}
\arguments{
\item{b}{(matrix) beta values}

\item{g}{(factor) grouping variable}
}
\value{
(matrix) mean of values belonging to same group
}
\description{
average beta values by group
}

