% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas_AOV.R
\name{betas_AOV}
\alias{betas_AOV}
\title{run ANOVA across groups}
\usage{
betas_AOV(betas, groupv, topX = 10000L, numCores = 8L)
}
\arguments{
\item{betas}{(matrix) probe-wise beta values}

\item{groupv}{(factor) group values, corresponding to columns of betas}

\item{topX}{(integer) limit to top variable probes. If NULL, uses all}

\item{numCores}{(integer) for parallel processing}
}
\value{
(list) topVar: probe IDs for top variable probes
}
\description{
run ANOVA across groups
}

