% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorByArray.R
\name{colorByArray}
\alias{colorByArray}
\title{color phenotype by position on array}
\usage{
colorByArray(v, arrayNum, palName = "Dark2", ttl = "Array layout", nr = 4,
  nc = 2, arrayPos = NULL)
}
\arguments{
\item{v}{(factor) variable of interest, one value per array}

\item{arrayNum}{(integer) which array will sample be on? same length as v}

\item{palName}{(char) RColorBrewer palette for colour-coding}

\item{ttl}{(char) plot title}

\item{nr}{(integer) num rows per array}

\item{nc}{(integer) num cols per array}

\item{arrayPos}{(integer) location within the array. if NULL, uses 
order in v}
}
\description{
color phenotype by position on array
}

