% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlProbes_oob.R
\name{controlProbes_oob}
\alias{controlProbes_oob}
\title{explore out of bound control probes}
\usage{
controlProbes_oob(rgSet, xlim = c(5, 17), sampNames = NULL)
}
\arguments{
\item{rgSet}{(RGSet) raw read-in values}

\item{xlim}{(numeric, length 2) acceptable range of intensities}

\item{sampNames}{(char) sample names for reference}
}
\description{
explore out of bound control probes
}
\details{
this script gets the address of the offending probes for 
communication with Illumina
}

