% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot_batch.R
\name{densityPlot_batch}
\alias{densityPlot_batch}
\title{density plot beta in chunks of samples}
\usage{
densityPlot_batch(rgset, chunkSize = 8, verbose = TRUE)
}
\arguments{
\item{rgset}{(RGChannelSet) raw experiment}

\item{chunkSize}{(integer) num samples to plot on each page}

\item{verbose}{(logical) print reassuring messages}
}
\value{
Nothing. Side effect of calling minfi::densityPlot() on each
subset of samples
}
\description{
density plot beta in chunks of samples
}

