% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmp_Fisher.R
\name{dmp_Fisher}
\alias{dmp_Fisher}
\title{run FET to ascertain overep. of sig probes}
\usage{
dmp_Fisher(fg_GR, bg_GR, targets)
}
\arguments{
\item{fg_GR}{(GRanges) foreground}

\item{bg_GR}{(GRanges) background}

\item{targets}{(data.frame or list) if char, path to bed files 
containing genomic regions (Type, description, File); 
if list assumes a list of GRanges}
}
\value{
(matrix) one row per data type. num_fg, num_bg, pct_fg, pct_bg, pvalue from FET
}
\description{
run FET to ascertain overep. of sig probes
}

