% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmp_pathwayGREAT.R
\name{dmp_pathwayGREAT}
\alias{dmp_pathwayGREAT}
\title{prepare files for manual pathway analysis by GREAT}
\usage{
dmp_pathwayGREAT(mset, dmpIn, geneFile, cutoff = 0.05, geneDomain = 50,
  outDir)
}
\arguments{
\item{mset}{(GenomicRatioSet)}

\item{dmpIn}{(DataFrame) output of minfi::dmpFinder()}

\item{geneFile}{(char) path to bed file with gene definitions}

\item{geneDomain}{(numeric) max distance (kb) from TSS at which a probe is considered
within the gene's domain. Used for GREAT analysis. Should be large enough to include
any probe that would be part of a gene's domain by GREAT's calculation. Should be small
enough that the resulting bed files can be uploaded to the GREAT server.
(http://bejerano.stanford.edu/great/public/html/)}

\item{outDir}{(char) path to output dir}
}
\description{
prepare files for manual pathway analysis by GREAT
}

