% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmp_pathwayORA.R
\name{dmp_pathwayORA}
\alias{dmp_pathwayORA}
\title{pathway overrepresentation by FET/binomial}
\usage{
dmp_pathwayORA(mset, thresh = 0.05, dmpIn, gene_GR, geneDomain = 5,
  pathwayFile, outDir, selMode = "pvalue", gmtFDRcutoff = 0.05,
  verbose = TRUE)
}
\arguments{
\item{mset}{(GenomicRatioSet)}

\item{dmpIn}{(DataFrame) output of dmp-wise analysis.}

\item{gene_GR}{(GRanges) gene bodies. Gene symbol should be in "name"
slot}

\item{geneDomain}{(integer) distance (kb) upstream from TSS
TSS that should be included in gene's domain. e.g.  Setting 
geneDomain=10 would result in [TSS-10kb,TES] being counted as a gene's
domain.}

\item{pathwayFile}{(char) path to gmt file with gene-sets}

\item{outDir}{(char) path to dir to write output files}

\item{selMode}{(char) pvalue|qvalue; if pvalue, foreground is probes
with p < 0.05; if qvalue, foreground is probes with Q < 0.05}

\item{gmtFDRcutoff}{(numeric 0-1) Q-value cutoff to use for writing gmt}

\item{verbose}{(logical) print messages}

\item{bg_GR}{(GRanges) background}
}
\value{
(data.frame) pathway statistics including Name, num probes 
overlapping in foreground (n_fg), in background (n_bg), pvalues for 
Hypergeometric (Hypergeom_p) and Binomial two-sided test (Binom_p).
and Benjamini-Hochberg adjusted Q values too (Hypergeom_Q, Binom_Q)
}
\description{
pathway overrepresentation by FET/binomial
}

