% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmp_pathwayORA_writeGMTFromRes.R
\name{dmp_pathwayORA_writeGMTFromRes}
\alias{dmp_pathwayORA_writeGMTFromRes}
\title{write GMT from results of pathway ORA}
\usage{
dmp_pathwayORA_writeGMTFromRes(pathwayFile, selPathways, fg_GR, gene_GR,
  geneDomain = 5, outFile)
}
\arguments{
\item{fg_GR}{(GRanges) foreground DMP probes. These will be used
to filter for genes that contribute to signal}

\item{gene_GR}{(GRanges) domains for each gene, may include upstream or
downstream of TSS.}

\item{geneDomain}{(integer) num kilobases to extend gene upstream of TSS}

\item{outFile}{(char) output file}

\item{pathwayRes}{(data.frame) table output of dmp_pathwayORA. 
columns are: Pathway, n_fg, n_bg, Hypergeom_p, Binomial_p,
Hypergeom_Q, Binom_Q, genes_fg}

\item{pathwayList}{(list) keys are pathways, values are member genes}
}
\description{
write GMT from results of pathway ORA
}
\details{
To be run when you have pathway-level p and Q values from
FET and binomial test, and you just want to generate the gmt to
visualize an EnrichmentMap
}

