% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmp_plotLocalBeta.R
\name{dmp_plotLocalBeta}
\alias{dmp_plotLocalBeta}
\title{plot betas in window around selected probes}
\usage{
dmp_plotLocalBeta(mset, selProbes, group = NULL, winWidth = 20000,
  writeMelted = FALSE, outPref = NULL)
}
\arguments{
\item{mset}{(MethylSet)}

\item{selProbes}{(char) probes to plot}

\item{winWidth}{(numeric) window width centered on selProbes}

\item{writeMelted}{(logical) if TRUE write melted matrix to file.
has probe name and sample info}

\item{outPref}{(char) prefix for output file}
}
\description{
plot betas in window around selected probes
}

