% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmp_plotLocalP.R
\name{dmp_plotLocalP}
\alias{dmp_plotLocalP}
\title{plot pvalue around hits, generate significant region list}
\usage{
dmp_plotLocalP(mset, dmpRes, selProbes, winWidth = 20000, FDR1 = 0.1,
  FDR2 = 0.3, minWin = 5000, buffer = 50, verbose = FALSE)
}
\arguments{
\item{mset}{(MethylSet)}

\item{dmpRes}{(data.frame) named data.frame with 'pval' column}

\item{selProbes}{(char) probes to plot}

\item{winWidth}{(numeric) window width centered on selProbes}

\item{FDR1}{(num 0-1) probes with higher significance, marked
in red in the plot}

\item{FDR2}{(num 0-1) larger than FDR1, probes with moderate
significance, marked in orange in the plot. Also used to 
win}

\item{minWin}{(integer) the minimum window width around a 
signal probe if no other moderate signal probes are present}

\item{buffer}{(integer) width by which to extend a region}

\item{plotMe}{(logical) if FALSE does not plot windows}
}
\value{
(data.frame) windows around significant probes
}
\description{
plot pvalue around hits, generate significant region list
}
\details{
A window of minWin is examined around each significant 
probe (selProbes). If it contains probes with FDR < FDR2, the
window is extended to the maximum distance of such a probe,
with a buffer of 10%.
If no such probes are present, the window has a minimum width
of minWin.
}

