% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmp_plotProbes_cont.R
\name{dmp_plotProbes_cont}
\alias{dmp_plotProbes_cont}
\title{plot probe effects}
\usage{
dmp_plotProbes_cont(mset, xvar = "age", selIdx, showSamp = 12L,
  setSeed = 42L, vals, ...)
}
\arguments{
\item{mset}{(GenomicRatioSet)}

\item{xvar}{(char) which pheno to plot on x-axis}

\item{selIdx}{(integer) probe indices. these probes will be plotted}

\item{showSamp}{(integer) max number of probes to show}

\item{setSeed}{(integer) RNG seed}

\item{vals}{(char) column name in elementMetadata(mset) to plot
as part of title}

\item{...}{params for plot() function}
}
\description{
plot probe effects
}

