% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmp_plotProbes_groups.R
\name{dmp_plotProbes_groups}
\alias{dmp_plotProbes_groups}
\title{plot betas by group, for selected probes}
\usage{
dmp_plotProbes_groups(mset, group = NULL, baseline_group = NULL,
  group2 = NULL, selProbes)
}
\arguments{
\item{mset}{(GenomicRatioSet)}

\item{group}{(char) which group variable to separate probes}

\item{baseline_group}{(char) which class in 'group' is the baseline?
function draws a reference line at the median for that value}

\item{group2}{(char; optional) split boxplots within a group using a 
second conditional variable. e.g. group="DX" and group2="SEX",
will show a primary split by diagnosis and secondary by sex. If NULL
not used}

\item{selProbes}{(char) probes to plot}

\item{...}{params for plot() function}
}
\description{
plot betas by group, for selected probes
}

