% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmpAge_2groups.R
\name{dmpAge_2groups}
\alias{dmpAge_2groups}
\title{compare <15 yo to >15yo. probe-wise analysis}
\usage{
dmpAge_2groups(pd, mset, outDir, runORA = FALSE, Qcutoff = 0.05, gene_GR,
  runGroups = "neuron")
}
\arguments{
\item{pd}{(data.frame) phenotype matrix}

\item{mset}{(MethylSet-Genomic)}

\item{outDir}{(char) path to dir where results should be written}

\item{runORA}{(logical) use gprofiler to run functional ORA}

\item{Qcutoff}{(numeric between 0 and 1) Q value for sig probes}

\item{gene_GR}{(GRanges) gene definitions, used only if runORA=TRUE}

\item{runGroups}{(char) which groups to run for? [neuron|glia]}
}
\value{
Nothing. Side effect of writing following tables/plots:
1) dmp_age_2groups_<tis>_<dt>: stats for significant probes. From dmpRun()
2) dmp_age_2groups_<tis>_<dt>.pdf: nominal p hist, qqplot, intercept hist.
from dmpRun()
3) dmp_age_<tis>_heatmap_<dt>: heatmap of sig probes.
4) age_2groups_<tis>_[ALL|UP|DOWN][prom|gene]_gProfileR_<dt>: gprofiler
output if runORA=TRUE
}
\description{
compare <15 yo to >15yo. probe-wise analysis
}

