% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationPie.R
\name{getLocationPie}
\alias{getLocationPie}
\title{pie chart of location breakdown}
\usage{
getLocationPie(gr, bedList, pal = "Spectral", verbose = TRUE)
}
\arguments{
\item{gr}{(GRanges) ranges of interest}

\item{bedList}{(data.frame) names of categories (Type) and path to bed 
files with ranges (File)}

\item{verbose}{(logical) print messages}
}
\description{
pie chart of location breakdown
}
\details{
subtractive. Starts with first category and moves down from
there. Does not guarantee that categories are mutually exclusive.
}

