% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbeDistr.R
\name{getProbeDistr}
\alias{getProbeDistr}
\title{get probe subset overlapping ROI. NOT strand-aware}
\usage{
getProbeDistr(query, subject, getSubset = FALSE)
}
\arguments{
\item{query}{(GRanges) probe locations}

\item{subject}{(char or GRanges) if char, vector of bed files to read in as 
GRanges and compute ol.}

\item{getSubset}{(logical) if TRUE, returns probes that overlap gene ranges}
}
\value{
probe breakdown
}
\description{
get probe subset overlapping ROI. NOT strand-aware
}

