% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{manhattan function, adapted from qqman}
\usage{
manhattan(x, chr = "CHR", bp = "BP", p = "P", snp = "SNP",
  col = c("gray10", "gray60"), chrlabs = NULL,
  suggestiveline = -log10(1e-05), genomewideline = -log10(5e-08),
  highlight = NULL, logp = TRUE, cex_hilite = 2, col_hilite = "red",
  cex = 0.7, ...)
}
\description{
manhattan function, adapted from qqman
}

