% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanPlot.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Manhattan plot of probe p results}
\usage{
manhattanPlot(mset, dmpTable)
}
\arguments{
\item{mset}{(MSet.genome)}

\item{dmpTable}{(data.frame) dmp results. Output of topTable()
function, must contain probe IDs as rownames, and the columns
"pval" (nominal p value) and "qval" (FDR corrected Q)}
}
\description{
Manhattan plot of probe p results
}

