% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAgeTrends.R
\name{plotAgeTrends}
\alias{plotAgeTrends}
\title{plot age effect for a set of genomic intervals}
\usage{
plotAgeTrends(mset, subject)
}
\arguments{
\item{mset}{(MethylSet-genomic)}

\item{subject}{(char or GRanges) see getProbeDistr}
}
\description{
plot age effect for a set of genomic intervals
}

