% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGene.R
\name{plotGene}
\alias{plotGene}
\title{plot methylation change with age for single genomic interval}
\usage{
plotGene(mset, gene_GR)
}
\description{
plot methylation change with age for single genomic interval
}

