% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSets.R
\name{plotGeneSets}
\alias{plotGeneSets}
\title{plot trends for gene sets}
\usage{
plotGeneSets(mset, gmt, gene_GR)
}
\arguments{
\item{mset}{(MethylSet-Genomic)}

\item{gmt}{(char) path to gmt file}

\item{gene_GR}{(GRanges)}
}
\description{
plot trends for gene sets
}

