% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion.R
\name{plotRegion}
\alias{plotRegion}
\title{plot methylation for a region}
\usage{
plotRegion(mset, rgn, gp, buffer = 1000, showGenes = TRUE, gene_GR)
}
\arguments{
\item{mset}{(MethylSet-Genomic)}

\item{rgn}{(GenomicRanges) range to show}

\item{gp}{(factor) value for each sample}

\item{buffer}{(integer) +/-bp to plot around rgn}

\item{showGenes}{(logical) show gene bodies?}

\item{gene_GR}{(Granges) ranges of genes, if showGenes=TRUE}
}
\description{
plot methylation for a region
}

