% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSelWin.R
\name{plotSelWin}
\alias{plotSelWin}
\title{plots a signal in windows of interest}
\usage{
plotSelWin(gr, cname, targetWin, refLine = 0, ...)
}
\arguments{
\item{gr}{(GRanges) signal of interest}

\item{cname}{(char) column with signal to plot}

\item{targetWin}{(GRanges) windows in which to plot the signal}

\item{refLine}{(numeric) yintercept for horizontal reference
line}

\item{...}{params for plot()}
}
\description{
plots a signal in windows of interest
}

