% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionDynamics.R
\name{regionDynamics}
\alias{regionDynamics}
\title{examine dynamics in regions of interest}
\usage{
regionDynamics(mset, target_GR, xvar = "age", showSamp, ttl = "", ...)
}
\arguments{
\item{mset}{(GenomicRatioSet)}

\item{target_GR}{(GRanges) regions of interest}

\item{xvar}{(char) pheno variable to use on x-axis}

\item{showSamp}{(integer) individual probes to show}

\item{ttl}{(char) string to put in plot title}

\item{...}{param for sample_nonParametric()}
}
\description{
examine dynamics in regions of interest
}
\details{
Finds subset of probes overlapping ROI. Then compares the rate
and direction of methylation change in this subset, relative to the 
backgound
}

