% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_nonParametric.R
\name{sample_nonParametric}
\alias{sample_nonParametric}
\title{get a pvalue by sampling with replacement from background}
\usage{
sample_nonParametric(fg, bg, R = 1000, FUN = median, type = "twotailed",
  showPlot = TRUE, numCores = 1L)
}
\arguments{
\item{fg}{(numeric) foreground set}

\item{bg}{(numeric vector) background set}

\item{FUN}{(function) function to compare}

\item{type}{(char) [twotailed|fg_greater|fg_less] type of tailed test}

\item{showPlot}{(logical) show real and null distributions}

\item{numCores}{(integer) for parallel processing}
}
\description{
get a pvalue by sampling with replacement from background
}

